[gaim-migrate @ 6037] Constification.
Also, what was I going--oh yeah, fixes a compile warning in oscar.c

Maybe I should be casting all of the 0.5s to ints?  Eh.

--HG--
extra : convert_revision : 24921ffda42a16569a5c49553dc02408dc6df52b

+++ src/protocols/oscar/oscar.c
@@ -3176,7 +3176,7 @@ static int gaim_parse_user_info(aim_session_t *sess, aim_frame_t *fr, ...) {
 			"%s\n"
 			"<hr>\n"),
 			info->sn, images(info->flags),
-			(info->warnlevel/10.0) + 0.5,
+			(int)((info->warnlevel/10.0) + 0.5),
 			onlinesince ? onlinesince : "",
 			membersince ? membersince : "",
 			idle ? idle : "");
